Implement a pool of hyper::Client
Provides a pool of hyper::Client workers for maximizing throughput on
multicore systems. A pool is created by passing it some configuration
including the number of clients to pool and how many sockets each client
should try and use.

Clients are stored in a list. Each request that is added starts at the
beginning of the list and tries to run on the first client that has
room. This has the effect of preferring loading up one event loop on a
single thread instead of spreading evenly across all workers.

    let mut config = Config::default();
    config.workers = 8;
    config.max_sockets = 10_000;

    let pool = Pool::new(&mut config);
    let (tx, rx) = mpsc::channel();

    for _ in 0..5 {
        pool.request(
            Url::parse("https://www.httpbin.org").unwrap(),
            Transaction::new(
                tx.clone(),
                Method::Get,
                Default::default(),
                None
            )
        ).expect("request ok");
    }

    assert_eq!(pool.clients[0].active_transactions(), 5);

    let mut received = 0;
    while let Ok(result) = rx.recv() {
        println!("got result: {:?}\n\n", result);

        received += 1;
        if received == 5 {
            break;
        }
    }