Add experimental new error handling
Clients are now restarted when one is detected to have been disconnected
(panicked).

The transaction counting has been improved to not use the Arc strong
count and instead use an atomic usize. This should be slightly better in
theory since the strong count uses sequential consistency and this only
requires acquire/release.