Change shutdown loop timeout to 100ms
The transaction timeout can be set to anything - I think in production
it's in the range of 60s. This results in shutdowns taking a very long
time, if you're not lucky enough to have no transactions in flight when
the shutdown is requested. I think it is OK to use a short polling loop
here because it is only used in the shutdown process. I tried adding
polling/async ability to the raii-counter library directly, and it would
require adding instrumentation overhead to each and every add/subtract
that we do. I think this is a better solution, as it requires no
overhead during the normal runtime process.